Ticket-Board-main-20241203T223757Z-001/Ticket-Board-main/requirements.txt
flask
requests
python-dotenv
PyYAML